- Move RubbishBin out of ComponentsPanel on to PreviewConcierge to get trashability when not in ComponentsPanel too - Add DeleteRoundedTiltedRight - and it's outline version — for better sidebar on the left coherence - Add BrokenLinkRounded icon to be used when "trashing" shared components
