+++ ui/app/src/components/Icons/BrokenLinkRounded.tsx
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2007-2019 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+import SvgIcon, { SvgIconProps } from '@material-ui/core/SvgIcon';
+
+export default function (props: SvgIconProps) {
+  return (
+    <SvgIcon {...props}>
+      <path d="M8.24163 12.6377C7.70018 12.7332 7.17865 12.368 7.08317 11.8266C6.9877 11.2851 7.35288 10.7636 7.89433 10.6681L10.8488 10.1472L11.1961 12.1168L8.24163 12.6377Z" />
+      <path d="M9.92103 16.4033C10.4625 16.3078 10.8277 15.7863 10.7322 15.2449C10.6367 14.7034 10.1152 14.3382 9.57373 14.4337L6.61931 14.9546C4.99399 15.2412 3.43045 14.146 3.14394 12.5212C2.85744 10.8963 3.95211 9.33238 5.57742 9.04579L8.53185 8.52485C9.0733 8.42938 9.43848 7.90784 9.34301 7.36639C9.24753 6.82494 8.726 6.45976 8.18455 6.55523L5.23013 7.07618C2.51229 7.55541 0.6951 10.1506 1.17433 12.8685C1.65356 15.5863 4.24877 17.4035 6.96661 16.9243L9.92103 16.4033Z" />
+      <path d="M14.8227 11.3627C15.3642 11.2672 15.8857 11.6324 15.9812 12.1739C16.0766 12.7153 15.7115 13.2368 15.17 13.3323L12.2156 13.8533L11.8683 11.8836L14.8227 11.3627Z" />
+      <path d="M13.1433 7.59712C12.6019 7.69259 12.2367 8.21413 12.3322 8.75558C12.4276 9.29703 12.9492 9.66221 13.4906 9.56674L16.445 9.04579C18.0703 8.75921 19.6339 9.85443 19.9204 11.4793C20.2069 13.1041 19.1122 14.6681 17.4869 14.9546L14.5325 15.4756C13.991 15.5711 13.6259 16.0926 13.7213 16.634C13.8168 17.1755 14.3383 17.5407 14.8798 17.4452L17.8342 16.9243C20.5521 16.445 22.3692 13.8498 21.89 11.132C21.4108 8.41414 18.8156 6.59695 16.0977 7.07618L13.1433 7.59712Z" />
+      <path d="M11.2888 2.74144C11.3008 2.46556 11.5342 2.25165 11.8101 2.26367C12.086 2.27568 12.2999 2.50906 12.2879 2.78494L12.194 4.94084C12.182 5.21672 11.9486 5.43062 11.6727 5.41861C11.3968 5.4066 11.1829 5.17322 11.1949 4.89733L11.2888 2.74144Z" />
+      <path d="M11.9642 20.9273C11.9406 21.2024 11.6985 21.4063 11.4233 21.3827C11.1482 21.3592 10.9443 21.117 10.9679 20.8419L11.1522 18.6918C11.1757 18.4167 11.4179 18.2128 11.693 18.2363C11.9682 18.2599 12.1721 18.5021 12.1485 18.7772L11.9642 20.9273Z" />
+      <path d="M8.36766 3.84267C8.24174 3.59691 8.3389 3.2956 8.58467 3.16969C8.83043 3.04377 9.13174 3.14093 9.25765 3.38669L10.2416 5.30724C10.3675 5.553 10.2704 5.85431 10.0246 5.98022C9.77885 6.10614 9.47755 6.00898 9.35163 5.76322L8.36766 3.84267Z" />
+      <path d="M14.9292 19.9498C15.0447 20.2006 14.9349 20.4976 14.6841 20.6131C14.4333 20.7286 14.1363 20.6188 14.0208 20.368L13.1184 18.4078C13.0029 18.157 13.1126 17.86 13.3634 17.7446C13.6143 17.6291 13.9112 17.7388 14.0267 17.9896L14.9292 19.9498Z" />
+      <path d="M13.8569 3.97398C14.005 3.74091 14.314 3.67203 14.547 3.82013C14.7801 3.96823 14.849 4.27723 14.7009 4.5103L13.5435 6.33163C13.3954 6.5647 13.0864 6.63358 12.8534 6.48548C12.6203 6.33738 12.5514 6.02838 12.6995 5.79531L13.8569 3.97398Z" />
+      <path d="M9.4503 19.5878C9.29254 19.8145 8.98093 19.8703 8.75428 19.7126C8.52764 19.5548 8.47179 19.2432 8.62955 19.0165L9.86234 17.2454C10.0201 17.0188 10.3317 16.9629 10.5584 17.1207C10.785 17.2784 10.8408 17.59 10.6831 17.8167L9.4503 19.5878Z" />
+    </SvgIcon>
+  );
+}
