+++ ui/app/src/components/Icons/DeleteRoundedTiltedRightOutline.tsx
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2007-2019 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+import SvgIcon, { SvgIconProps } from '@material-ui/core/SvgIcon';
+
+export default function DeleteRoundedTiltedRight(props: SvgIconProps) {
+  return (
+    <SvgIcon {...props}>
+      <path d="M5.41647 18.6542C4.41953 18.1893 3.98421 16.9932 4.44909 15.9963L8.67527 6.93322C9.14015 5.93628 10.3362 5.50096 11.3331 5.96584L18.5836 9.34679C19.5805 9.81167 20.0158 11.0077 19.551 12.0046L15.3248 21.0677C14.8599 22.0647 13.6639 22.5 12.6669 22.0351L5.41647 18.6542ZM16.832 10.7368L11.3942 8.20108C10.8957 7.96864 10.2977 8.1863 10.0653 8.68477L6.68432 15.9352C6.45188 16.4337 6.66954 17.0317 7.16801 17.2642L12.6059 19.7999C13.1043 20.0323 13.7023 19.8146 13.9348 19.3162L17.3157 12.0657C17.5482 11.5672 17.3305 10.9692 16.832 10.7368Z"/>
+      <path d="M6.35243 5.01906L18.1701 2.93528C18.7118 2.83977 19.0768 2.31847 18.9813 1.77682C18.8858 1.23518 18.3645 0.870156 17.8228 0.965663L15.3608 1.39978L14.5383 0.82386C14.3298 0.677851 14.0546 0.614671 13.7986 0.659819L9.68209 1.38567C9.42604 1.43082 9.18909 1.58429 9.04308 1.79282L8.46715 2.61532L6.00514 3.04944C5.46349 3.14495 5.09847 3.66625 5.19398 4.2079C5.28948 4.74954 5.81079 5.11456 6.35243 5.01906Z" />
+    </SvgIcon>
+  );
+}
